// src/app/core/types/type.ts
export interface DadosCadastroUsuario {
  nomeCompleto: string;
  nomeUsuario: string;
  email: string;
  senha: string;
}

export interface DadosListagemUsuario {
  email: string;
  nomeCompleto: string;
  nomeUsuario: string;
}

export interface PessoaUsuaria {
 nomeCompleto: string;
  nomeUsuario: string;
  email: string;
  senha: string;
}